data-collection/requirements.txt
arrow==0.15.4    
boto3==1.10.41   
google-api-python-client==1.7.11
google-auth==1.9.0     
google-auth-httplib2==0.0.3     
google-auth-oauthlib==0.4.1     
